Reduce rescan overheads in TID [Range] Scan
Until how the TIDs to scan have been recalculated after every rescan of
both TID Scan and TID Range Scan.  Here we adjust things so that the
calculated list of TIDs or calculated range of TIDs is only marked as
needing recalculated whenever a parameter has changed that might require
a recalculation.