Make truncate_useless_pathkeys() consider WindowFuncs
truncate_useless_pathkeys() seems to have neglected to account for
PathKeys that might be useful for WindowClause evaluation.  Modify it so
that it properly accounts for that.

To make this work, we also modify pathkeys_useful_for_ordering() so it
checks sort_pathkeys rather than query_pathkeys.  This was a problem
because nothing was checking sort_pathkeys, so if query_pathkeys got set
to another value for some other operation, then nothing would account
for PathKeys useful for the ORDER BY clause.