Cleanup INFINITY code in float.h
The INFINITY macro is always defined per C99 standard, so this should
mean we can now get rid of the workaround code for when that macro isn't
defined.

Also, delete the #pragma code which was disabling a compiler warning in
MSVC.  There was a comment explaining why the #pragma was placed outside
the function body to work around a MSVC compiler bug, but the link
explaining that was dead, as reported by jian he.

Reported-by: jian he <jian.universality@gmail.com>
Author: David Rowley <dgrowleyml@gmail.com>
Discussion: https://postgr.es/m/CACJufxGARYETnNwtCK7QC0zE_7gq-tfN0mME=gT5rTNtC=VSHQ@mail.gmail.com