Tidyup truncate_useless_pathkeys() function
This removes a few static functions and replaces them with 2 functions
which aim to be more reusable.  The upper planner's pathkey requirements
can be simplified down to operations which require pathkeys in the same
order as the pathkeys for the given operation, and operations which can
make use of a Path's useful pathkeys in any order.

Here we also add some short-circuiting paths to
truncate_useless_pathkeys().  Since the function will only return a
trimmed down pathkey list if *no* operation finds some of the pathkeys
useful, we might as well exit early as soon as we discover that some
operation finds *all* of the pathkeys will be useful.  I've also tried
to place the pathkey checks roughly in order of cheapest-first in the
hopes that the more expensive checks might get skipped once in a while
when the function discovers all pathkeys are useful.

In passing fixup has_useful_pathkeys() as it seems to have grown a
redundant check for group_pathkeys.  This isn't needed as
standard_qp_callback will set query_pathkeys if there's any requirement
to have group_pathkeys.