Fix BRIN 32-bit counter wrap issue with huge tables
A BlockNumber (32-bit) might not be large enough to add bo_pagesPerRange
to when the table contains close to 2^32 pages.  At worst, this could
result in an infinite loop for power-of-2 pagesPerRange, or inefficient
BRIN index scans for non-power-of-2 pagesPerRange.

Author: sunil s <sunilfeb26@gmail.com>
Discussion: https://postgr.es/m/CAOG6S4-tGksTQhVzJM19NzLYAHusXsK2HmADPZzGQcfZABsvpA@mail.gmail.com